add date and user in reviews

+++ main.py
@@ -2,8 +2,11 @@ import argparse
 import csv
 import requests
 from bs4 import BeautifulSoup
+import os
+from datetime import datetime
 from restaurant_class import Restaurant
 from restaurant_info import restaurant_info
+from get_reviews import get_all_reviews
 
 
 LINK = 'https://www.opentable.com/m/best-restaurants-in-america-for-2017/'
@@ -15,22 +18,33 @@ def write_csv(name, links):
     :param links: list of links of all restaurants
     :return: None
     """
-    print('Writing reviews.csv')
-    with open('reviews.csv', 'w') as f:
-        f.write('Name,Place,Comment,Overall,Food,Service,Ambience,Date,Vip\n')
+    if not os.path.exists("scrap_date.txt"):
+        old_date='1900-01-01 00:00:00.000000'
+        print('Writing reviews.csv')
+        with open('reviews.csv', 'w') as f:
+            f.write('Name,Place,Comment,Overall,Food,Service,Ambience,Date,Vip,User,N_rev\n')
+        scrap_date = datetime.strptime(old_date.split('.')[0], "%Y-%m-%d %H:%M:%S")
+    else:
+        with open('scrap_date.txt', 'r') as f:
+            text_date = f.readline().split('.')[0]
+            scrap_date = datetime.strptime(text_date, "%Y-%m-%d %H:%M:%S")
+
+    print('Appending scrape results')
     for n_res in range(len(links)):
         print('Writing', name[n_res], n_res, "of", len(links))
-        temp_res = Restaurant(links[n_res])
         try:
-            (places, comments, overall, food, service, ambience, dates, vips) = temp_res.get_all_reviews()
+            (places, comments, overall, food, service, ambience, dates, vips, users, n_revs) = get_all_reviews(links[n_res], scrap_date)
         except:
             print("No reviews found for:", name[n_res])
         else:
             with open('reviews.csv', 'a') as csvfile:
                 spamwriter = csv.writer(csvfile, delimiter=',', quoting=csv.QUOTE_MINIMAL)
-                for n_rev in range(len(places)):
-                    row = [name[n_res], places[n_rev], comments[n_rev], overall[n_rev], food[n_rev], service[n_rev], ambience[n_rev], dates[n_rev], vips[n_rev]]
+                for i_rev in range(len(places)):
+                    row = [name[n_res], places[i_rev], comments[i_rev], overall[i_rev], food[i_rev], service[i_rev], ambience[i_rev], dates[i_rev], vips[i_rev], users[i_rev], n_revs[i_rev]]
                     spamwriter.writerow(row)
+
+    with open('scrap_date.txt', 'w') as f:
+        f.write(str(datetime.now()))
 
     #Writing restaurants
     print('Writing 100restaurants.csv')
@@ -66,8 +80,6 @@ def main():
     if args.w:
         print("writting csv")
         (rest_links, restaurants) = get_links_and_names()
-        # rest_links = ['https://www.opentable.com/1770-house', 'https://www.opentable.com/arethusa-al-tavolo']
-        # restaurants = ['1770 House', 'Arethusa al tavolo']
         write_csv(restaurants, rest_links)
